Cleanup and fix tests
